requirements.txt
PyQt5
pyqtwebengine
jinja2
pyyaml
